% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omdb_api_calls.R
\name{add_key_to_renviron}
\alias{add_key_to_renviron}
\title{Add your api key to .Renviron}
\usage{
add_key_to_renviron(key = NULL, location = "~/.Renviron")
}
\arguments{
\item{key}{keep NULL to search or add manualy}

\item{location}{what file do you want to write to?  defaults to ~/.Renviron}
}
\description{
This is a convenient function that writes
your key to the .Renviron file that is read in
at start. You dont have to give the key in, if
you use keep the key NULL it will ask for it in
the commandline (keeping it out of the logs and .history).
}
\details{
Of course you can also go to the file and add it yourself.
the file should have a line that reads:
OMDB_KEY=key

where you replace your key with your key.
}
