% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichIMDB.R
\name{enrichIMDB}
\alias{enrichIMDB}
\title{IMDB information enricher.}
\usage{
enrichIMDB(df)
}
\arguments{
\item{df}{Name of dataframe you want to add info to. NEEDS an imdbID variable.}
}
\description{
This function downloads episode information into a dataframe. 
Use it to create a dataframe with extra information about every epidode.
it downloads runtime, director, writer, actors, plot and imdb votes.
}
\examples{
\dontrun{ 
enrichIMDB("IMDB")
}
}
\keyword{enrich}
\keyword{imdb,}

