% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/movies.R
\name{getMovies}
\alias{getMovies}
\title{movies}
\usage{
getMovies(tables = c("movies.list.gz", "actors.list.gz", "actresses.list.gz",
  "directors.list.gz"), temp.dir = tempdir())
}
\arguments{
\item{tables}{a character vector of files from IMDB to download. The default is
movies, actors, actresses, and directors. There are 49 total files available on IMDB.}

\item{temp.dir}{a directory where you want to store the download files.}
}
\value{
An integer indicating the number of files present in \code{temp.dir} whose
filenames match those on IMDB.
}
\description{
Download the raw data files from IMDB
}
\examples{
if (require(RPostgreSQL) & require(dplyr)) {
 # must have pre-existing database "airlines"
 db <- src_postgres(host = "localhost", user="postgres", password="postgres", dbname = "imdb")
}
\dontrun{
  getMovies(temp.dir = "~/dumps/imdb")
}
}

