% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-imdb.R
\name{IMDB.getActorsFromMovie}
\alias{IMDB.getActorsFromMovie}
\title{IMDB.getActorsFromMovie}
\usage{
IMDB.getActorsFromMovie(ttid, return.cols = NULL, imdb = imdb.data)
}
\arguments{
\item{ttid}{Movie identifier `ttid`}

\item{return.cols}{list of columns you want to display, by default all}

\item{imdb}{If you want to apply this function to a different dataframe, it's possible

dim(imdb.data$all.movies.actors.characters);     # 901324      4}
}
\value{
dataframe of results
}
\description{
IMDB.getActorsFromMovie
}
