% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdbMovie.R
\name{imdbMovies}
\alias{imdbMovies}
\title{Download information about a movie from imdb}
\usage{
imdbMovies(moviename)
}
\arguments{
\item{moviename}{Type the name of the movie}
}
\description{
Download information about a movie from imdb
}
