% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdbTitleId.R
\name{imdbTitleId}
\alias{imdbTitleId}
\title{Download information about a title from imdb by imdb ID}
\usage{
imdbTitleId(titleid, key = NULL)
}
\arguments{
\item{titleid}{The id, f.i.: "tt0120382"}

\item{key}{Enter your API key. If you don't have one yet, you can obtain one here: http://www.omdbapi.com/apikey.aspx}
}
\description{
Download information about a title from imdb by imdb ID
}
