% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/07-bo_bt_ratio-vs-other-vars.R
\name{plot_bo_bt_ratio_vs_others}
\alias{plot_bo_bt_ratio_vs_others}
\title{Generate exploratory plots for boxoffice/budget ratio vs. other variables}
\usage{
plot_bo_bt_ratio_vs_others()
}
\value{
15 ggplot2 objects
}
\description{
Generate exploratory plots for boxoffice/budget ratio vs. other variables
}

