% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/06-boxoffice-vs-other-vars.R
\name{plot_bo_vs_others}
\alias{plot_bo_vs_others}
\title{Generate descriptive plots for genres}
\usage{
plot_bo_vs_others()
}
\value{
3 ggplot2 objects
}
\description{
Generate descriptive plots for genres
}

