% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-films-n-votes.R
\name{plot_films_n_votes}
\alias{plot_films_n_votes}
\title{Generate descriptive plots for film and vote counts}
\usage{
plot_films_n_votes()
}
\value{
6 ggplot2 objects
}
\description{
Generate descriptive plots for film and vote counts
}
