% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-lengths-n-mpaa.R
\name{plot_length_n_mpaa}
\alias{plot_length_n_mpaa}
\title{Generate descriptive plots for ratings and votes per rating}
\usage{
plot_length_n_mpaa()
}
\value{
2 ggplot2 objects
}
\description{
Generate descriptive plots for ratings and votes per rating
}
