% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/03-ratings-n-votes-per-rating.R
\name{plot_ratings}
\alias{plot_ratings}
\title{Generate descriptive plots for ratings and votes per rating}
\usage{
plot_ratings()
}
\value{
3 ggplot2 objects
}
\description{
Generate descriptive plots for ratings and votes per rating
}

