% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{push}
\usage{
push(db, temp.dir, path.to.imdbpy2sql = NULL)
}
\arguments{
\item{db}{a \code{dplyr} \code{src} or a \code{DBI} connection}

\item{temp.dir}{a directory containing the IMDB files you with to process}

\item{path.to.imdbpy2sql}{a path to the IMDB2SQL Python script provided by
IMDBPy. If NULL -- the default -- will attempt to find it using \code{\link{findimdbpy2sql}}.}
}
\description{
a utility function to push a data frame to a DB.
}
\examples{
\dontrun{
if (require(RMySQL) & require(dplyr)) {
 # must have pre-existing database "imdb"
 db <- src_mysql(host = "localhost", user="bbaumer", password="fakepass", dbname = "imdb")
}

cache <- "~/dumps/imdb"
  getMovies(temp.dir = cache)
  push(db, temp.dir = cache)
}
}

