% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tosviolationmessage.R
\name{ToS_message}
\alias{ToS_message}
\alias{tos_violation}
\alias{termsofservice}
\alias{tosviolation}
\title{ToS violation message}
\usage{
ToS_message()
}
\description{
It has come to my attention that the use of the omdbapi might violate the terms of service
that imdb has set up. According to \url{http://www.imdb.com/conditions} scraping is forbidden. 
I did not scrape IMDB but since we are using the omdb-api, we sort of violate the terms of service 
indirectly. This is just a heads up. 
 
If you would like to stop this message, add a line to your .Renviron file: IMDBTOGGLE = "off". If you do I presume
you have read the message.
}
