% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\name{etl_extract.etl_imdb}
\alias{etl_extract.etl_imdb}
\alias{etl_load.etl_imdb}
\alias{etl_load_data}
\alias{etl_load_data.src_mysql}
\title{movies}
\source{
IMDB: \url{ftp://ftp.fu-berlin.de/pub/misc/movies/database/}

IMDbPy: \url{http://imdbpy.sourceforge.net/}
}
\usage{
\method{etl_extract}{etl_imdb}(obj, tables = c("movies", "actors",
  "actresses", "directors"), all.tables = FALSE, ...)

\method{etl_load}{etl_imdb}(obj, schema = TRUE, path_to_imdbpy2sql = NULL,
  password = "", ...)

etl_load_data(obj, ...)

\method{etl_load_data}{src_mysql}(obj, ...)
}
\arguments{
\item{obj}{an \code{\link{etl}} object}

\item{tables}{a character vector of files from IMDB to download. The default is
movies, actors, actresses, and directors. These four files alone will occupy
more than 500 MB of disk space. There are 49 total files available 
on IMDB. See \url{ftp://ftp.fu-berlin.de/pub/misc/movies/database/} for the
complete list.}

\item{all.tables}{a logical indicating whether you want to download all of 
the tables. Default is \code{FALSE}.}

\item{...}{arguments passed to methods}

\item{schema}{Either a logical, a filename pointing to SQL schema file, or a character
vector containing the schema itself. If \code{schema = TRUE}, then the built-in
schema will be used. Note
that the flavor of SQL in this file must match the type of the source. That is,
if your object is of type \code{\link[dplyr]{src_mysql}}, then make sure that
the schema you specify here is written in MySQL (and not PostgreSQL). Please
note that SQL syntax is not, in general, completely portable. Use with caution, as this may
clobber any existing data you have in an existing database.}

\item{path_to_imdbpy2sql}{a path to the IMDB2SQL Python script provided by
IMDBPy. If NULL -- the default -- will attempt to find it using \code{\link{findimdbpy2sql}}.}

\item{password}{Must re-enter password unless your password is blank. The real
password will not be shown in messages.}
}
\description{
Download the raw data files from IMDB
}
\details{
For best performance, set the MySQL default collation to \code{utf8_unicode_ci}.
See the IMDbPy2sql documentation at 
\url{http://imdbpy.sourceforge.net/docs/README.sqldb.txt} for more details. 

Please be aware that IMDB contains information about *all* types of movies.
}
\examples{
# Connect using default RSQLite database
imdb <- etl("imdb")

# Connect using pre-configured PostgreSQL database
\dontrun{
 if (require(RPostgreSQL)) {
   # must have pre-existing database "imdb"
   db <- src_postgres(host = "localhost", user="postgres", password="postgres", dbname = "imdb")
  }
  imdb <- etl("imdb", db = db, dir = "~/dumps/imdb/")
  imdb \%>\%
    etl_extract(tables = "movies") \%>\%
    etl_load()
}
\dontrun{
 if (require(RMySQL)) {
   # must have pre-existing database "imdb"
   db <- src_mysql(default.file = path.expand("~/.my.cnf"), 
                   group = "scidb", user = NULL, password = NULL, dbname = "imdb")
  }
  imdb <- etl("imdb", db = db, dir = "~/dumps/imdb/")
  imdb \%>\%
    etl_extract(tables = "movies") \%>\%
    etl_load()
    
  movies <- imdb \%>\%
    tbl("title") 
  movies \%>\%
    filter(title == 'star wars')
    
  people <- imdb \%>\%
    tbl("name") 
  roles <- imdb \%>\%
    tbl("cast_info") 
  movies \%>\%
    inner_join(cast_info, by = c("id" = "movie_id")) \%>\%
    inner_join(people, by = c("person_id" = "id")) \%>\%
    filter(title == 'star wars') \%>\%
    filter(production_year == 1977) \%>\%
    arrange(nr_order)
  
}
}

