% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findimdbpy2sql}
\alias{findimdbpy2sql}
\title{Find the path to the imdbpy2sql.py script}
\usage{
findimdbpy2sql(temp_dir = tempdir())
}
\arguments{
\item{temp_dir}{directory where you want the IMDBPy script to reside}
}
\value{
a path to the unzipped, executable script
}
\description{
A utility function to locate the \code{imdbpy2sql} Python script
}
\examples{
script <- findimdbpy2sql()
}

