% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imdbMovie.R
\name{imdbMovies}
\alias{imdbMovies}
\title{Download information about a movie from imdb}
\usage{
imdbMovies(moviename, key = NULL)
}
\arguments{
\item{moviename}{Type the name of the movie}

\item{key}{Enter your API key. If you don't have one yet, you can obtain one here: http://www.omdbapi.com/apikey.aspx}
}
\description{
Download information about a movie from imdb
}
