% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/06-boxoffice-vs-other-vars.R
\name{plot_bo_vs_others}
\alias{plot_bo_vs_others}
\title{Generate exploratory plots for boxoffice vs. other variables}
\usage{
plot_bo_vs_others()
}
\value{
14 ggplot2 objects
}
\description{
Generate exploratory plots for boxoffice vs. other variables
}

