% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-budget-n-boxoffice.R
\name{plot_budget_n_boxoffice}
\alias{plot_budget_n_boxoffice}
\title{Generate descriptive plots for budget and boxoffice}
\usage{
plot_budget_n_boxoffice()
}
\value{
8 ggplot2 plots
}
\description{
Generate descriptive plots for budget and boxoffice
}
