% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/03-ratings-n-votes-per-rating.R
\name{plot_ratings}
\alias{plot_ratings}
\title{Generate some descriptive plots for budget and boxoffice over the years}
\usage{
plot_ratings()
}
\value{
ggplot2 plots
}
\description{
Generate some descriptive plots for budget and boxoffice over the years
}

